+++ src/bios/memory.c
@@ -5,7 +5,7 @@
 
 // Look up the memory region corresponding to an address. Returns NULL if the
 // address is not mapped to a known memory region.
-MemoryRegion* GetMemoryRegion(struct BIOSState* bios, uint16_t address) {
+MemoryRegion* GetMemoryRegion(struct BIOSState* bios, uint32_t address) {
   for (uint8_t i = 0; i < MemoryRegionsLength(&bios->memory_regions); ++i) {
     MemoryRegion* region = MemoryRegionsGet(&bios->memory_regions, i);
     if (address >= region->start && address < region->start + region->size) {
@@ -16,7 +16,7 @@ MemoryRegion* GetMemoryRegion(struct BIOSState* bios, uint16_t address) {
 }
 
 // Read a byte from a logical memory address.
-uint8_t ReadLogicalMemoryByte(struct BIOSState* bios, uint16_t address) {
+uint8_t ReadLogicalMemoryByte(struct BIOSState* bios, uint32_t address) {
   MemoryRegion* region = GetMemoryRegion(bios, address);
   if (!region || !region->read_memory_byte) {
     return 0xFF;
@@ -25,7 +25,7 @@ uint8_t ReadLogicalMemoryByte(struct BIOSState* bios, uint16_t address) {
 }
 
 // Read a word from a logical memory address.
-uint16_t ReadLogicalMemoryWord(struct BIOSState* bios, uint16_t address) {
+uint16_t ReadLogicalMemoryWord(struct BIOSState* bios, uint32_t address) {
   uint8_t low_byte = ReadLogicalMemoryByte(bios, address);
   uint8_t high_byte = ReadLogicalMemoryByte(bios, address + 1);
   return (high_byte << 8) | low_byte;
@@ -33,7 +33,7 @@ uint16_t ReadLogicalMemoryWord(struct BIOSState* bios, uint16_t address) {
 
 // Write a byte to a logical memory address.
 void WriteLogicalMemoryByte(
-    struct BIOSState* bios, uint16_t address, uint8_t value) {
+    struct BIOSState* bios, uint32_t address, uint8_t value) {
   MemoryRegion* region = GetMemoryRegion(bios, address);
   if (!region || !region->write_memory_byte) {
     return;
@@ -43,7 +43,7 @@ void WriteLogicalMemoryByte(
 
 // Write a word to a logical memory address.
 void WriteLogicalMemoryWord(
-    struct BIOSState* bios, uint16_t address, uint16_t value) {
+    struct BIOSState* bios, uint32_t address, uint16_t value) {
   WriteLogicalMemoryByte(bios, address, value & 0xFF);
   WriteLogicalMemoryByte(bios, address + 1, (value >> 8) & 0xFF);
 }
